NestorBot/NestorApp/requirements.txt
slackclient==2.9.1
slackeventsapi==2.2.1
pymongo==3.9.0
